[KOGITO-1474] Fix issue Kogito service cannot be scaled to 0 (#220)

+++ bddframework/steps/kogitojobsservice.go
@@ -40,7 +40,7 @@ func (data *Data) kogitoJobsServiceHasPodsRunningWithinMinutes(pods, timeoutInMi
 }
 
 func (data *Data) scaleKogitoJobsServiceToPodsWithinMinutes(nbPods, timeoutInMin int) error {
-	err := framework.SetKogitoJobsServiceReplicas(data.Namespace, nbPods)
+	err := framework.SetKogitoJobsServiceReplicas(data.Namespace, int32(nbPods))
 	if err != nil {
 		return err
 	}
